[EI-167] [web] refactor: Update numerical guidance components
