+++ web/__tests__/components/numerical-guidance/indicator-list.test.tsx
@@ -2,7 +2,7 @@ import { act, render, renderHook, screen, waitFor } from '@testing-library/react
 import userEvent from '@testing-library/user-event';
 import { SWRProviderWithoutCache } from '@/app/store/querys/swr-provider';
 import { resetMockDB } from '@/app/mocks/db.mock';
-import IndicatorList from '@/app/ui/components/numerical-guidance/molecule/indicator-list';
+import IndicatorList from '@/app/ui/components/numerical-guidance/indicator/indicator-list';
 import { useNumericalGuidanceStore } from '@/app/store/stores/numerical-guidance.store';
 import { resetAllStore } from '@/app/store/stores/reset-store';
