+++ web/app/ui/components/numerical-guidance/indicator/advanced-indicators-chart.tsx
