+++ web/app/ui/components/numerical-guidance/indicator/indicator-dialog-menu.tsx
