+++ web/app/ui/components/numerical-guidance/indicator/indicator-list-item.tsx
