+++ web/app/ui/components/numerical-guidance/indicator/indicators-chart.tsx
