+++ web/app/ui/pages/workspace/dialog-container.tsx
@@ -1,4 +1,4 @@
-import IndicatorDialogMenu from '../../components/numerical-guidance/molecule/indicator-dialog-menu';
+import IndicatorDialogMenu from '../../components/numerical-guidance/indicator/indicator-dialog-menu';
 import MetadataDeleteDialog from '../../components/numerical-guidance/indicator-board-metadata/metadata-delete-dialog';
 import MetadataDialogMenu from '../../components/numerical-guidance/indicator-board-metadata/metadata-dialog-menu/metadata-dialog-menu';
